% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cortex.R
\docType{data}
\name{cortex}
\alias{cortex}
\title{Macaque Visual Cortex Network}
\format{
an igraph object with "gml" format
}
\usage{
data("cortex")
}
\description{
A graph describing the macaque visual cortex network.
Nodes are neocortical areas, 25 of them are participated in visual function in the macaque,
and 7 of which are associated with them
}
\examples{
data("cortex")
print(cortex)
}
\references{
D.J. Felleman and D.C. van Essen, "Distributed hierarchical processing in the primate cerebral cortex." Cerebral Cortex 1(1), 1-47 (1991).
}
\keyword{datasets}
