% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINNA.R
\name{print.calculate.centralities}
\alias{print.calculate.centralities}
\title{Print computed centrality measures results into a file}
\usage{
\method{print}{calculate.centralities}(x, file = NULL)
}
\arguments{
\item{x}{a list containing the centrality measure values}

\item{file}{A character string naming the .pdf file to print into.
If NULL the result would be printed to the exist directory.(default=NULL)}
}
\value{
Print out  \code{ \link[CINNA]{calculate_centralities}}function will be saved in the given directory.
}
\description{
This function prints all centrality measure results into a file
}
\author{
Minoo Ashtiani, Mohieddin Jafari
}
