% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINNA.R
\name{print.visualize.graph}
\alias{print.visualize.graph}
\title{Print visualized based on centrality values graph}
\usage{
\method{print}{visualize.graph}(x, computed.centrality.value = NULL,
  centrality.type = "Degree Centrality", file = NULL)
}
\arguments{
\item{x}{an igraph object}

\item{computed.centrality.value}{A vector containing the values of calculated centrality measure for each node(default=NULL)}

\item{centrality.type}{The type of centrality which should be calculated(default="Degree Centrality")}

\item{file}{A character string naming the .pdf file to print into. If NULL the result would be printed to the exist directory.(default=NULL)}
}
\description{
This function prints visualized based on centrality values graph.
}
\author{
Minoo Ashtiani, Mohieddin Jafari
}
