% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINNA.R
\name{print.visualize.heatmap}
\alias{print.visualize.heatmap}
\title{Print centralities' heatmap}
\usage{
\method{print}{visualize.heatmap}(x, scale = TRUE, file = NULL)
}
\arguments{
\item{x}{a list indicating calculated centrality measures}

\item{scale}{Whether the centrality values should be scaled or not(default=TRUE)}

\item{file}{A character string naming the .pdf file to print into. If NULL
the result would be printed to the exist directory.(default=NULL)}
}
\description{
This function prints the heatmap plot
}
\author{
Minoo Ashtiani, Mohieddin Jafari
}
