% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINNA.R
\name{print.visualize.pair.correlation}
\alias{print.visualize.pair.correlation}
\title{Print pairwise correlation of centrality measure and histogram plot}
\usage{
\method{print}{visualize.pair.correlation}(x, y, scale = TRUE, file = NULL)
}
\arguments{
\item{x}{a vector containing a centrality measure}

\item{y}{a vector containing another centrality measure}

\item{scale}{Whether the centrality values should be scaled or not(default=TRUE)}

\item{file}{A character string naming the .pdf file to print into. If NULL the result would be printed to the exist directory.}
}
\description{
This function prints pairwise correlation of centrality measure and histogram plot.
}
\author{
Minoo Ashtiani, Mohieddin Jafari
}
