% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINNA.R
\name{print_visualize_dendrogram}
\alias{print_visualize_dendrogram}
\title{Print dendrogram plot of a graph}
\usage{
print_visualize_dendrogram(x, centrality.type = "Degree Centrality",
  computed_centrality_value = NULL, k = 4, file = NULL)
}
\arguments{
\item{x}{an igraph object}

\item{centrality.type}{The type of centrality which should be calculated(default="Degree Centrality")}

\item{computed_centrality_value}{A vector containing the values of calculated
centrality measure for each node(default=NULL)}

\item{k}{number of clusters}

\item{file}{A character string naming the .pdf file to print into.
If NULL the result would be printed to the exist directory.(default=NULL)}
}
\value{
The resulted plot of \code{ \link[CINNA]{visualize_dendrogram}}function will be saved in the given directory.#' @importFrom igraph alpha.centrality
}
\description{
This function prints dendrogram plot of a graph based on predefined
centrality measures.
}
\author{
Minoo Ashtiani, Mohieddin Jafari
}
