% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINNA.R
\name{print_visualize_graph}
\alias{print_visualize_graph}
\title{Print visualized based on centrality values graph}
\usage{
print_visualize_graph(x, computed_centrality_value = NULL,
  centrality.type = "Degree Centrality", file = NULL)
}
\arguments{
\item{x}{an igraph object}

\item{computed_centrality_value}{A vector containing the values of calculated centrality measure for each node(default=NULL)}

\item{centrality.type}{The type of centrality which should be calculated(default="Degree Centrality")}

\item{file}{A character string naming the .pdf file to print into. If NULL the result would be printed to the exist directory.(default=NULL)}
}
\value{
The resulted plot of \code{ \link[CINNA]{visualize_graph}}function will be saved in the given directory.#' @importFrom igraph alpha.centrality
}
\description{
This function prints visualized based on centrality values graph.
}
\author{
Minoo Ashtiani, Mohieddin Jafari
}
