% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINNA.R
\name{proper_centralities}
\alias{proper_centralities}
\title{Proper centrality measure representation}
\usage{
proper_centralities(x)
}
\arguments{
\item{x}{an igraph object}
}
\value{
a list including the name of centrality measures which are suitable for the input graph
}
\description{
This function indicates proper centrality measures of an
igraph object based on the network topology
}
\details{
This function represents a list including the names of centrality measures which are applicable
for the input graph based on the topology
}
\examples{

data("zachary")
proper_centralities(zachary)

}
\seealso{
\code{\link[CINNA]{calculate_centralities}}
}
\author{
Minoo Ashtiani, Mohieddin Jafari
}
