% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhesus.R
\docType{data}
\name{rhesus}
\alias{rhesus}
\title{Moreno Rhesus Network}
\format{
an igraph object with "gml" format
}
\usage{
rhesus
}
\description{
A directed graph including observed grooming episodes between
free ranging rhesus macaques (Macaca mulatta) in Cayo Santiago
during a two month period in 1963. Cayo Santiago is an island off the coast of Puerto Rico,
which also is named as Isla de los monos (Island of the monkeys).
A node indicates a monkey and a directed edge in which
a rhesus macaque groomed another rhesus macaque.
The weights of edges demonstrates how often this behaviour
was seen.
}
\examples{
data("rhesus")
print(rhesus)
}
\references{
Rhesus network dataset -- KONECT, October 2016.

DS Sade. Sociometrics of macaca mulatta I. linkages and cliques in grooming matrices. Folia Primatologica, 18(3-4):196--223, 1972.
}
\keyword{datasets}
