% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINNA.R
\name{summary.calculate.centralities}
\alias{summary.calculate.centralities}
\title{Summarize centrality measure calculation results}
\usage{
\method{summary}{calculate.centralities}(x)
}
\arguments{
\item{x}{centrality measure calculation results}
}
\value{
a list concluding summary results for each centrality measure value

The result values of \code{ \link[CINNA]{calculate_centralities}}function will be saved in the given directory.#' @importFrom igraph alpha.centrality
}
\description{
This function computes  minimum, first quater, median,
mean, third qarter and maximum values of computed centrality measures.
}
\author{
Minoo Ashtiani, Mohieddin Jafari
}
