% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINNA.R
\name{summary.graph.extract.components}
\alias{summary.graph.extract.components}
\title{Summarize component extraction of a graph}
\usage{
\method{summary}{graph.extract.components}(x, directed = TRUE,
  bipartite_proj = FALSE, num_proj = 1)
}
\arguments{
\item{x}{An igraph or a network object}

\item{directed}{a boolean constant, Whether to create a directed graph(default=TRUE)}

\item{bipartite_proj}{Whether the bipartite network must be projected or not(default=FALSE)}

\item{num_proj}{A number which shows the number of projects especifically for bipartite graphs.(default=1)}
}
\value{
The result values of \code{ \link[CINNA]{graph_extract_components}}function will be saved in the given directory.#' @importFrom igraph alpha.centrality
}
\description{
This function summarizes all components of the input which can be an "igraph" object or a "network" object
}
\author{
Minoo Ashtiani, Mohieddin Jafari
}
