% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINNA.R
\name{summary_pca_centralities}
\alias{summary_pca_centralities}
\title{Summarize PCA result related to centrality measures}
\usage{
summary_pca_centralities(x, scale.unit = TRUE, ncp = 5)
}
\arguments{
\item{x}{a list containg the computed centrality values}

\item{scale.unit}{a boolean constant, whether data should be scaled to unit
variance(default=TRUE)}

\item{ncp}{number of dimensions in final results (default=5)}
}
\value{
The result values of \code{ \link[CINNA]{pca_centralities}}function will be saved in the given directory.#' @importFrom igraph alpha.centrality
}
\description{
This function summarizes PCA result related to centrality measures.
}
\author{
Minoo Ashtiani, Mehdi Mirzaie, Mohieddin Jafari
}
