% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINNA.R
\name{visualize.graph}
\alias{visualize.graph}
\title{Graph visualization based on a specific centrality measure}
\usage{
visualize.graph(x, computed.centrality.value = NULL,
  centrality.type = "Degree Centrality")
}
\arguments{
\item{x}{an igraph object}

\item{computed.centrality.value}{A vector containing the values of calculated centrality measure for each node.}

\item{centrality.type}{The type of centrality which should be calculated.}
}
\value{
a plot illustrating the graph
}
\description{
This function demonstrates the input graph in which the size of nodes
indicates calculated centrality value.
}
\details{
This function represents the graph in which size of nodes are based on computed centrality value. If the values of wanted centrality
measure were computed then by placing them in computed.centrality.value argument to use it for drawing the plot. Otherwise, by only giving the
name of favorite centrality measure in centrality.type argument, this function will calculate it and  then demonstrates the corresponding graph.
}
\author{
Minoo Ashtiani, Mohieddin Jafari
}
