% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINNA.R
\name{visualize.heatmap}
\alias{visualize.heatmap}
\title{Heatmap between  centrality values}
\usage{
visualize.heatmap(x, scale = TRUE)
}
\arguments{
\item{x}{a list indicating calculated centrality measures}

\item{scale}{Whether the centrality values should be scaled or not}
}
\value{
The correlation plot
}
\description{
This function draws heatmap between pair of centrality measures
}
\details{
This function illustrates the heatmap plot of computed centrality measures.
}
\author{
Minoo Ashtiani, Mohieddin Jafari
}
