% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINNA.R
\name{visualize_association}
\alias{visualize_association}
\title{Pairwise association plot between centrality measures}
\usage{
visualize_association(x, y, scale = TRUE)
}
\arguments{
\item{x}{a vector containing a centrality measure as independent variable}

\item{y}{a vector containing a centrality measure as dependent variable}

\item{scale}{Whether the centrality values should be scaled or not}
}
\value{
The regression plot, and the values resulted by the regression process.
}
\description{
This function computes regression between pair of centrality measures
to show more details of association among them.
}
\details{
This function applies regression analysis on two different centrality values in order to find out the corresponding association
between them. Regression analysis is a kind of statitiscal method for approximation the association
between variables.It asserts that the value of dependent variable changes when the
value of independent variable varies.
}
\references{
CHAMBERS, & M., J. (1992). Statistical Models in S. Wadsworth. Pacific Grove, California. Retrieved from

Wilkinson, G. N., & Rogers, C. E. (1973). Symbolic Description of Factorial Models for Analysis of Variance. Applied Statistics, 22(3), 392.
}
\author{
Minoo Ashtiani, Mohieddin Jafari
}
