% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINNA.R
\name{visualize_heatmap}
\alias{visualize_heatmap}
\title{Heatmap plot between centrality measures}
\usage{
visualize_heatmap(x, scale = TRUE)
}
\arguments{
\item{x}{a list indicating calculated centrality measures}

\item{scale}{Whether the centrality values should be scaled or not}
}
\value{
The function generates a heatmap plot between pairs of centrality measures.
The function takes a list x as input, which contains the computed centrality measures.
If the scale parameter is set to TRUE, the centrality values will be scaled before plotting the heatmap.
The function creates a heatmap plot, where each cell represents the correlation between a pair of centrality measures.
The color of each cell indicates the strength and direction of the correlation.
The function returns the heatmap plot.
}
\description{
This function draws heatmap between pair of centrality measures
}
\details{
This function illustrates the heatmap plot of computed centrality measures.
}
\author{
Minoo Ashtiani, Mehdi Mirzaie, Mohieddin Jafari
}
