% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINNA.R
\name{visualize_pair_correlation}
\alias{visualize_pair_correlation}
\title{Pairwise correlation between two centrality values}
\usage{
visualize_pair_correlation(x, y, scale = TRUE)
}
\arguments{
\item{x}{a vector containing a centrality measure}

\item{y}{a vector containing another centrality measure}

\item{scale}{Whether the centrality values should be scaled or not}
}
\value{
The correlation plot
}
\description{
This function computes and plots correlation between pair of centrality
 measures and histogram plots.
}
\details{
This function illustrates the correlation value between two centrality measures and their
corresponding scatterplot and histograms.
}
\references{
Emerson, J. W., Green, W. A., Schloerke, B., Crowley, J., Cook, D., Hofmann, H., & Wickham, H. (2013). The Generalized Pairs Plot. Journal of Computational and Graphical Statistics, 22(1), 79–91.
}
\seealso{
\code{\link[GGally]{ggpairs}}
}
\author{
Minoo Ashtiani, Mohieddin Jafari
}
