% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zachary.R
\docType{data}
\name{zachary}
\alias{zachary}
\title{Zachary Karate Club Network}
\format{
an igraph object with "gml" format
}
\usage{
zachary
}
\description{
A graph describing friendships among members of a university
karate club. Includes metadata for faction membership after a social
partition.
}
\examples{
data("zachary")
print(zachary)
}
\references{
W. W. Zachary, "An information flow model for conflict and fission in small groups." Journal of Anthropological Research 33, 452-473 (1977).
}
\keyword{datasets}
