% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINNA.R
\name{proper_centralities}
\alias{proper_centralities}
\title{Proper centrality measure representation}
\usage{
proper_centralities(x)
}
\arguments{
\item{x}{An igraph object.}
}
\value{
A list including the names of centrality measures that are suitable for the input graph.
}
\description{
This function indicates the proper centrality measures of an igraph object
based on the network topology.
}
\details{
This function returns a list including the names of centrality measures that are applicable
for the input graph based on its topology.
}
\examples{

data("zachary")
proper_centralities(zachary)

}
\seealso{
\code{\link[CINNA]{calculate_centralities}}
}
\author{
Minoo Ashtiani, Mehdi Mirzaie, Mohieddin Jafari
}
