% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINNA.R
\name{summary.graph.extract.components}
\alias{summary.graph.extract.components}
\title{Summarize component extraction of a graph}
\usage{
\method{summary}{graph.extract.components}(x, directed = TRUE,
  bipartite.proj = FALSE, num.proj = 1)
}
\arguments{
\item{x}{An igraph or a network object}

\item{directed}{a boolean constant, Whether to create a directed graph(default=TRUE)}

\item{bipartite.proj}{Whether the bipartite network must be projected or not(default=FALSE)}

\item{num.proj}{A number which shows the number of projects especifically for bipartite graphs.(default=1)}
}
\value{
a list including summary of all componets
}
\description{
This function summarizes all components of the input which can be an "igraph" object or a "network" object
}
\author{
Minoo Ashtiani, Mohieddin Jafari
}
