% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CINNA.R
\name{visualize_pair_correlation}
\alias{visualize_pair_correlation}
\title{Pairwise correlation plot between two centrality measures}
\usage{
visualize_pair_correlation(x, y, scale = TRUE)
}
\arguments{
\item{x}{a vector containing a centrality measure}

\item{y}{a vector containing another centrality measure}

\item{scale}{Whether the centrality values should be scaled or not}
}
\value{
The function computes and plots the correlation between two centrality measures.
        The function takes two vectors, `x` and `y`, representing the centrality measures to be correlated.
        If the `scale` parameter is set to `TRUE`, the centrality values will be scaled before computing the correlation.
        The function creates a correlation plot, which includes a scatterplot of the centrality values and histograms of each measure.
        The correlation coefficient between the two measures is displayed on the plot.
        The function returns the correlation plot.
}
\description{
This function computes and plots correlation between pair of centrality
 measures and histogram plots.
}
\details{
This function illustrates the correlation value between two centrality measures and their
corresponding scatterplot and histograms.
}
\references{
Emerson, J. W., Green, W. A., Schloerke, B., Crowley, J., Cook, D., Hofmann, H., & Wickham, H. (2013). The Generalized Pairs Plot. Journal of Computational and Graphical Statistics, 22(1), 79–91.
}
\seealso{
\code{\link[GGally]{ggpairs}}
}
\author{
Minoo Ashtiani, Mehdi Mirzaie, Mohieddin Jafari
}
